+++ Sources/ProjectSpec/Dependency.swift
@@ -5,7 +5,7 @@ public struct Dependency: Equatable {
     public static let removeHeadersDefault = true
     public static let implicitDefault = false
     public static let weakLinkDefault = false
-    public static let platformDefault: Platform = .all
+    public static let platformFilterDefault: PlatformFilter = .all
 
     public var type: DependencyType
     public var reference: String
@@ -15,8 +15,8 @@ public struct Dependency: Equatable {
     public var link: Bool?
     public var implicit: Bool = implicitDefault
     public var weakLink: Bool = weakLinkDefault
-    public var platform: Platform = platformDefault
-    public var conditionalPlatforms: Set<ProjectSpec.Platform>?
+    public var platformFilter: PlatformFilter = platformFilterDefault
+    public var conditionalPlatforms: Set<Platform>?
 
     public init(
         type: DependencyType,
@@ -26,8 +26,8 @@ public struct Dependency: Equatable {
         link: Bool? = nil,
         implicit: Bool = implicitDefault,
         weakLink: Bool = weakLinkDefault,
-        platform: Platform = platformDefault,
-        conditionalPlatforms: Set<ProjectSpec.Platform>? = nil
+        platformFilter: PlatformFilter = platformFilterDefault,
+        conditionalPlatforms: Set<Platform>? = nil
     ) {
         self.type = type
         self.reference = reference
@@ -36,11 +36,11 @@ public struct Dependency: Equatable {
         self.link = link
         self.implicit = implicit
         self.weakLink = weakLink
-        self.platform = platform
+        self.platformFilter = platformFilter
         self.conditionalPlatforms = conditionalPlatforms
     }
     
-    public enum Platform: String, Equatable {
+    public enum PlatformFilter: String, Equatable {
         case all
         case iOS
         case macOS
@@ -126,10 +126,10 @@ extension Dependency: JSONObjectConvertible {
             weakLink = bool
         }
         
-        if let platformString: String = jsonDictionary.json(atKeyPath: "platform"), let platform = Platform(rawValue: platformString) {
-            self.platform = platform
+        if let platformFilterString: String = jsonDictionary.json(atKeyPath: "platformFilter"), let platformFilter = PlatformFilter(rawValue: platformFilterString) {
+            self.platformFilter = platformFilter
         } else {
-            self.platform = .all
+            self.platformFilter = .all
         }
 
         if let conditionalPlatforms: [ProjectSpec.Platform] = jsonDictionary.json(atKeyPath: "conditionalPlatforms") {
@@ -144,7 +144,7 @@ extension Dependency: JSONEncodable {
             "embed": embed,
             "codeSign": codeSign,
             "link": link,
-            "conditionalPlatforms": conditionalPlatforms.map(Array.init)?.map(\.rawValue)
+            "conditionalPlatforms": conditionalPlatforms?.map(\.rawValue).sorted()
         ]
 
         if removeHeaders != Dependency.removeHeadersDefault {